+++ event-service/src/main/java/com/example/eventservice/domain/repository/event/query/OrderQuery.java
@@ -7,7 +7,8 @@ import com.querydsl.core.types.OrderSpecifier;
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.example.eventservice.entity.event.QCulturalEvent.culturalEvent;
+import static com.example.eventservice.domain.entity.event.QCulturalEvent.culturalEvent;
+
 
 public class OrderQuery {
